Move to last (dev) version of Jubako.
Now Jubako entry store creator takes an iterator on entries.
The values of the entries will be add by jubako itself to value stores
and such, so no need to handle that.

We use the crate internment to store the mimetypes as they are mostly
duplicated and interning them avoid unecessary copies.